dom/tests/mochitest/bugs/test_bug393974.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=393974
-->
<head>
  <title>Test for Bug 393974</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=393974">Mozilla Bug 393974</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

function test()
{
  // Create a tree walker with a filter which creates a cycle...
  var tw = document.createTreeWalker(document, NodeFilter.SHOW_ALL,
                                     function(n)
                                     {
                                       // force the closure to contain the
                                       // global object, in case a future
                                       // optimization might minimize the
                                       // function's captured environment
                                       if ("foo" + window == "fooPIRATES!")
                                         return NodeFilter.FILTER_ACCEPT;
                                       return NodeFilter.FILTER_REJECT;
                                     });

  // That should have been enough to create a leak, but we should do at least
  // a couple tests while we're here so that this document doesn't show up as
  // having no tests pass *or* fail.

  ok(tw.firstChild() === null, "shouldn't be a first child");
  ok(tw.currentNode === document, "should be unchanged");

  ok(tw.lastChild() === null, "shouldn't be a last child");
  ok(tw.currentNode === document, "should be unchanged");

  ok(tw.nextNode() === null, "shouldn't be a next node");
  ok(tw.currentNode === document, "should be unchanged");

  ok(tw.nextSibling() === null, "shouldn't be a next sibling");
  ok(tw.currentNode === document, "should be unchanged");

  ok(tw.parentNode() === null, "shouldn't be a parent node");
  ok(tw.currentNode === document, "should be unchanged");

  ok(tw.previousNode() === null, "shouldn't be a previous node");
  ok(tw.currentNode === document, "should be unchanged");

  ok(tw.previousSibling() === null, "shouldn't be a previous sibling");
  ok(tw.currentNode === document, "should be unchanged");
}

addLoadEvent(test);
addLoadEvent(SimpleTest.finish);
</script>
</pre>
</body>
</html>
